% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_twitter_posting.R
\name{tweet_posting}
\alias{tweet_posting}
\title{Function for tweeting
This function allows you to post tweets from R.}
\usage{
tweet_posting(url = "https://www.ulpgc.es/noticias")
}
\description{
Function for tweeting
This function allows you to post tweets from R.
}
\examples{
tweet_posting()

}
