% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_twitter_connection.R
\name{twitter_connection}
\alias{twitter_connection}
\title{Function for connecting to Twitter account
This function allows you to connect the twitter account to R.}
\usage{
twitter_connection()
}
\description{
Function for connecting to Twitter account
This function allows you to connect the twitter account to R.
}
\examples{
twitter_connection()

}
