% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_web_scrapping.R
\name{ulpgc_scrap}
\alias{ulpgc_scrap}
\title{Function for scrapping data from the ULPGC webpage
This function scraps the last news from the ULPGC news section page}
\usage{
ulpgc_scrap(url = "https://www.ulpgc.es/noticias")
}
\arguments{
\item{url}{URL from the ULPGC news section}
}
\value{
String with last news
}
\description{
Function for scrapping data from the ULPGC webpage
This function scraps the last news from the ULPGC news section page
}
\examples{
ulpgc_scrap()

}
